+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGELOG
 
-- UNRELEASED
+- 2024-07-05: 2.20.0
 
   - Dropped support for Ruby 3.0
+  - Add `Logster::Logger#subscribe` which allows subscribing to log events from `Logster::Logger`.
 
 - 2024-03-12: 2.19.1